fix: convert LoE section to vue component
